requirements.txt
black
django
django-extensions
django-crispy-forms
django-simple-menu